frontend(rights): move createTransaction function to FireQuery
