+++ libs/utils/src/lib/firequery/types.ts
@@ -26,3 +26,47 @@ export function createQuery<T>(path: string): Query<T> {
 export function isQueryLike<T>(query: QueryInput<T>): query is QueryLike<T> {
   return typeof query !== 'string'
 }
+
+////////////////
+// DOC RIGHTS //
+////////////////
+
+export interface OrgDocRights {
+  id: string;
+  canCreate: boolean;
+  canRead: boolean;
+  canUpdate: boolean;
+  canDelete: boolean;
+  isAdmin: boolean;
+}
+
+export interface SharedDocRights {
+  id: string;
+  canCreate: [];
+  canRead: [];
+  canUpdate: [];
+  canDelete: [];
+  admins: [];
+}
+
+export function initializeOwnerDocRights (id: string) {
+  return {
+    id,
+    canCreate: true,
+    canRead: true,
+    canUpdate: true,
+    canDelete: true,
+    isAdmin: true,
+  } as OrgDocRights;
+}
+
+export function initializeSharedDocRights (id: string) {
+  return {
+    id,
+    canCreate: [],
+    canRead: [],
+    canUpdate: [],
+    canDelete: [],
+    admins: [],
+  } as SharedDocRights;
+}
